% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguityDict.R
\name{clearAmbiguity}
\alias{clearAmbiguity}
\title{Clear up user defined ambiguity dictionary}
\usage{
clearAmbiguity()
}
\description{
Clear up user defined ambiguity dictionary.
}
\examples{
\dontrun{
clearAmbiguity()
}

}

