% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{clearStopwords}
\alias{clearStopwords}
\title{Clear up stopwords.}
\usage{
clearStopwords()
}
\value{
No results.
}
\description{
Clear up stopwords.
}
\examples{
\dontrun{
insertStopwords(c("of", "at"))
showStopwords()
clearStopwords()
showStopwords()
}
}
\seealso{
\code{\link{insertStopwords}},\code{\link{showStopwords}}
}

