% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{insertStopwords}
\alias{insertStopwords}
\title{Insert stopwords.}
\usage{
insertStopwords(stopwords)
}
\arguments{
\item{stopwords}{A character vector.}
}
\value{
No results.
}
\description{
Insert stopwords from a character vector.
}
\examples{
\dontrun{
insertStopwords(c("of", "at"))
showStopwords()
clearStopwords()
}
}
\seealso{
\code{\link{showStopWords}},\code{\link{clearStopWords}}
}

