% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguityDict.R
\name{loadAmbiguity}
\alias{loadAmbiguity}
\title{Load user defined ambiguity dictionary}
\usage{
loadAmbiguity(path)
}
\arguments{
\item{path}{charater, file path of the dictionary. The suffix of the
dictionary should be ".dic". There should be only one word per line.
Every line should be formated like "ChineseWords\tabFirstSegmentWord\tabPOS\tab
SecondSegmentWord\tabPOS...".}
}
\value{
No results.
}
\description{
Load user defined ambiguity dictionary by dictionary path.
}
\examples{
\dontrun{
dic.path = system.file("library/ambiguity.dic", package = "ransj")
loadAmbiguity(dic.path)
}
}

