% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userDict.R
\name{loadDict}
\alias{loadDict}
\title{Load user defined dictionary}
\usage{
loadDict(path)
}
\arguments{
\item{path}{charater, file path of the dictionary, or folder path to load all
the .dic files. The suffix of the dictionary should be ".dic". There should be
only one word per line. Every line should be formated like "Word\tabPOS\tabFrequency".}
}
\value{
No results.
}
\description{
Load user defined dictionary by dictionary path.
}
\examples{
\dontrun{
dic.path = system.file("library/default.dic", package = "ransj")
loadDict(dic.path)
}
}

