% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ransj.options.R
\name{ransj.options}
\alias{ransj.options}
\title{Set options of segmentation.}
\usage{
ransj.options(...)
}
\arguments{
\item{...}{\code{isNameRec},\code{isNumRec},\code{isQuantifierRec} can be defined by the way as \code{\link{options}}.}
}
\value{
No results.
}
\description{
Three options can be set to contol the process of the segmentation.
\code{isNameRec} means whether to recognize person names. Default is TRUE.
\code{isNumRec} means whether to recognize numbers. Default is TRUE.
\code{isQuantifierRec} means whether to combine the numbers and the quantifier. Default is TRUE.
}

