% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Segment Chinese sentence into words.}
\usage{
segment(str, method = "nlp", nature = FALSE, stopwords = FALSE,
  naturesInclude = NULL, naturesRemove = NULL, nosymbol = TRUE,
  returnType = "tm")
}
\arguments{
\item{str}{A character of chinese sentence to be segmented.}

\item{method}{A chrarcter to select one segment method from:
\itemize{
 \item{"nlp": }{NlpAnalysis of ansj. The Most Accurate segment method in ansj. User can define dictionary.
Numbers, person names, organition names and new words recognition
are supported in the method.}
 \item{"dic": }{DicAnalysis of ansj. User defined dictionary will have a higher priority.}
 \item{"base":}{BaseAnalysis of ansj.Sentences will be segment into very short words.
User defined Dictionary will not be used by this method. Numbers recognition is
included in this method. Person names, organization names and new words
recognition are not supported in this method.}
 \item{"to": }{ToAnalysis of ansj. More accurate than "base" method. User defined
dictionary will be used by the method. Numbers and person names recognition is
included in this method. Organization names and new words
recognition are not supported in this method.}
 \item{"index": }{IndexAnalysis of ansj. This method aims at producing search index.}
 \item{"fastIndex": }{FastIndexAnalysis of ansj. Faster than "index" method.}
}}

\item{nature}{logical. Whether to tag the POS. Default = FALSE.}

\item{stopwords}{logical. Whether to remove stopwords from the result.
Default = FALSE. Use \code{\link{insertStopwords}} first before you set
parameter \code{stopwords=TRUE}}

\item{naturesInclude}{Character vector. The natures to be take back from the result.
Default = NULL.}

\item{naturesRemove}{Character vector. The natures to be removed from the result.
Default = NULL.Note that if naturesInclude and naturesRemove are both not null,
naturesRemove will be ignored.}

\item{returnType}{A character from \code{c("vector", "tm")}. "vector" means
that the result returned will be a vector containing the words. "tm" means
that the result returned will be a character in which each word is separated
by space.}
}
\value{
A character or character vector depends on the parameter returnType.
}
\description{
A wrapper to ansj chinese words segmentation package. This function
support all segmentation methods of ansj: NlpAnalysis, ToAnalysis, BaseAnalysis,
DicAnalysis, IndexAnalysis, FastIndexAnalysis. More details about these methods
refers to https://github.com/NLPchina/ansj_seg
}

