% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{showStopwords}
\alias{showStopwords}
\title{Show stopwords.}
\usage{
showStopwords()
}
\value{
A character vector or character(0). Stopwords has been inserted.
}
\description{
Show stopwords.
}
\examples{
\dontrun{
insertStopwords(c("of", "at"))
showStopwords()
}
}
\seealso{
\code{\link{insertStopwords}},\code{\link{clearStopwords}}
}

